update templates in order to use translation strings
note the use of `{!!` instead of `{{` to echo the strings. It allows the strings to make use of html, including entities. This is only allowed if you have full control of the values you want to display. !!! Never display a value coming from elsewhere (i.e. the user, through a form) using those blade constructs !!!